[chore #166715682] add signup and signin tests
