--- test/signin.js
@@ -1,108 +1,0 @@
-import { describe, it } from 'mocha';
-import chai from 'chai';
-import chaiHttp from 'chai-http';
-import app from '../server/server';
-
-chai.use(chaiHttp);
-chai.should();
-
-describe('signin', () => {
-  it('user should be able to signin', (done) => {
-    const user = {
-      email: 'chris@gmail.com',
-      password: 'Asdfg1',
-    };
-    chai.request(app)
-      .post('/api/v1/auth/signin')
-      .send(user)
-      .end((err, res) => {
-        res.should.have.status(200);
-        res.should.be.an('object');
-        res.body.should.have.property('status').eql(200);
-        res.body.should.have.property('data');
-        done();
-      });
-  });
-
-  it('user should not be able to signin when the email is not registered', (done) => {
-    const user = {
-      email: 'afhdsfd@gmail.com',
-      password: 'Asdfg1',
-    };
-    chai.request(app)
-      .post('/api/v1/auth/signin')
-      .send(user)
-      .end((err, res) => {
-        res.should.have.status(404);
-        res.should.be.an('object');
-        res.body.should.have.property('status').eql(404);
-        res.body.should.have.property('error');
-        done();
-      });
-  });
-
-  it('user should not be able to signin when the password length is greater than 12', (done) => {
-    const user = {
-      email: 'chris@gmail.com',
-      password: 'Aaaaaaaaaaaaaa',
-    };
-    chai.request(app)
-      .post('/api/v1/auth/signin')
-      .send(user)
-      .end((err, res) => {
-        res.should.have.status(400);
-        res.should.be.an('object');
-        res.body.should.have.property('status').eql(400);
-        res.body.should.have.property('error');
-        done();
-      });
-  });
-  it('user should not be able to signin when there is an empty field', (done) => {
-    const user = {
-      email: '',
-      password: 'Asdfg1',
-    };
-    chai.request(app)
-      .post('/api/v1/auth/signin')
-      .send(user)
-      .end((err, res) => {
-        res.should.have.status(400);
-        res.should.be.an('object');
-        res.body.should.have.property('status').eql(400);
-        res.body.should.have.property('error');
-        done();
-      });
-  });
-  it('user should not be able to signin when the password length is less than 6', (done) => {
-    const user = {
-      email: 'chris@gmail.com',
-      password: 'Aaa',
-    };
-    chai.request(app)
-      .post('/api/v1/auth/signin')
-      .send(user)
-      .end((err, res) => {
-        res.should.have.status(400);
-        res.should.be.an('object');
-        res.body.should.have.property('status').eql(400);
-        res.body.should.have.property('error');
-        done();
-      });
-  });
-  it('user should not be able to signin when the password is incorrect', (done) => {
-    const user = {
-      email: 'chris@gmail.com',
-      password: 'aaaaaaaaa',
-    };
-    chai.request(app)
-      .post('/api/v1/auth/signin')
-      .send(user)
-      .end((err, res) => {
-        res.should.have.status(400);
-        res.should.be.an('object');
-        res.body.should.have.property('status').eql(400);
-        res.body.should.have.property('error');
-        done();
-      });
-  });
-});
